feat: added dynamic colour to thermometer ico
Added function to allow for the temperature in the bottom right corner to change in response to the temperature being selected. 

Added a realistic temperature range of 15 degrees celcius to 32 degrees celsius; this is for display and demonstration only. Numbers displayed for
temperature values are dummy values.

Adjusted colour used as background for the card in the bottom right corner of the climate page to allow for the current temperature colour to be viewed more easily in both light and dark mode.